refactor: move get_grid_case import statements to their respective modules
