+++ src/api/__init__.py
@@ -1,6 +1,5 @@
 
 from api.models import GridCase, GridCaseModel, ReconfigurationOutput
-from api.grid_cases import get_grid_case
 from api.ray_utils import (
     where_am_i,
     init_ray,
@@ -13,7 +12,6 @@ __all__ = [
     "GridCase",
     "GridCaseModel",
     "ReconfigurationOutput",
-    "get_grid_case",
     "where_am_i",
     "init_ray",
     "shutdown_ray",
